Make Fireworks and Together into the Adapter format
